simplify(S04b): rehome Agent helpers out of workquery.go
Pure same-package move: DrainTimeoutDuration and the nine
session-capacity helpers (EffectiveMax/MinActiveSessions, Supports*,
UsesCanonicalSingletonPoolIdentity, HasUnlimitedSessionCapacity,
ResolvedMaxActiveSessions) go to a new session_capacity.go beside the
other session-concern files; EffectiveDefaultSlingFormula joins the
formula-name accessors in config.go. workquery.go now holds only query
builders and resolvers. No behavior change, no caller edits.
